store: Rename Upsert -> Replace and add UpdateConfiguration to interface
The UpdateConfiguration is a bit more costly as it may require two write
operations: one for pushing attributes to the list and one to prune
duplicates (if any). Therefore it might not be a good idea to expose
this function to the external API (PATCH) at this point.

changelog: none

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>